% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DKdata-package.R
\docType{package}
\name{DKdata}
\alias{DKdata}
\alias{DKdata-package}
\title{Collection of DKdata}
\description{
Collection of Danish datasets
}
\details{
\tabular{ll}{ Package: \tab DKdata\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2012-03-29\cr License: \tab GPL-2\cr } % ~~ An overview of
how to use the package, including the most important ~~ % ~~ functions ~~
}
\author{
Claus Ekstrom \email{claus@rprimer.dk}\cr Maintainer: Claus Ekstrom
\email{claus@rprimer.dk}
}
\keyword{package}

