% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DKdata-package.R
\docType{data}
\name{precip}
\alias{precip}
\title{Precipitation in Copenhagen}
\format{A data frame with 51132 observations on the following 4 variables.
  \describe{
    \item{\code{year}}{a numeric vector}
    \item{\code{month}}{a numeric vector}
    \item{\code{day}}{a numeric vector}
    \item{\code{precip}}{a numeric vector (in mm)}
  }}
\source{
\url{http://www.ecad.eu}
}
\description{
Information on daily precipitation in Botanisk Have, Copenhagen, Denmark, from 1873 to 2013.
}
\details{
24-hour summed precipitation for Botanisk Have (Botanical Gardens), Copenhagen, coordinates:
55.68N, 12.58E, from 1873 to 2013.
}
\examples{

data(precip)
with(earthquakes, place[which.max(mag)])

}
\references{
Klein Tank, A.M.G. and Coauthors, 2002. Daily dataset of 20th-century surface air temperature
and precipitation series for the European Climate Assessment.  Int. J. of Climatol
}
\keyword{datasets}

