% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DKdata-package.R
\docType{data}
\name{sentiment}
\alias{sentiment}
\title{Danish sentiment wordlist}
\format{A data frame with 3552 rows and 2 variables:
\describe{
  \item{\code{word}}{character vector containing the words and phrases}
  \item{\code{score}}{integer vector with the sentiment score}
}}
\source{
Data collected by Finn Årup Nielsen and adapted from https://github.com/fnielsen
}
\usage{
sentiment
}
\description{
A wordlist of Danish words and phrases and their corresponding sentiment score (on a scale from -5 to 5)
}
\examples{

data(sentiment)
head(sentiment)

}
\keyword{datasets}
