% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DKdata-package.R
\docType{data}
\name{spurious}
\alias{spurious}
\title{Spurious Correlations}
\format{A data frame with 15 observations on the following 12 variables.
\describe{
\item{\code{Aar}}{The year}
\item{\code{Biografer}}{The number of movie theatres in Denmark}
\item{\code{Vaskemaskiner}}{The percentage of homes with a washing machine}
\item{\code{Skilsmisser}}{The number of divorces}
\item{\samp{Storebælt}}{Investments related to Storebæltsforbindelen (mio DKK)}
\item{\samp{Levendefødte}}{The number of live children born}
\item{\code{Udvandrede}}{The number of emigrants}
\item{\code{tvspots}}{The turn-over on tv commercials (1000 DKK).}
\item{\code{Stog}}{Transportation by S-trains (1000 individuals)}
\item{\samp{Årsmiddeltemperatur}}{The average yearly temperature}
\item{\code{Peter}}{The number of newborns given the name Peter}
\item{\samp{Brøndby}}{The ranking of the Brøndby football club in the football league}
}}
\source{
Data collected from Statistics Denmark, Wikipedia, and various web pages
}
\description{
Time series from 2001 to 2015 for a number of random (mostly) Danish statistics
to illustrate spurious correlations,
}
\examples{

data(spurious)

cor(spurious)


}
\keyword{datasets}
