% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{A vector of common Danish stop words for text mining}
\format{An object of class \code{character} of length 145.}
\usage{
stopwords
}
\description{
A vector of common Danish stop words for text mining
}
\author{
Claus Ekstrøm \email{claus@rprimer.dk}
}
\references{
Custom list by author
}
\keyword{datasets}

