% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DKdata-package.R
\docType{data}
\name{stopwords2}
\alias{stopwords2}
\title{Updated Danish stopword wordlist}
\format{A tibble with 177 rows and 1 variable:
\describe{
  \item{\code{word}}{character vector containing the stopwords}
}}
\source{
Data updated from stopwords with missing obvious stopwords by Claus Ekstrøm
}
\usage{
stopwords2
}
\description{
An updated wordlist of Danish stopwords. Based on a combination of the 120 most common Danish words, the stopwords data frame, and various other sources. Note that the word 'ikke' is part of the data and it may be necessary to remove that if the list is used in combination with sentiment analysis.
}
\examples{

data(stopwords2)
head(stopwords2)

}
\keyword{datasets}
