% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DKdata-package.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Danish stopword wordlist}
\format{A tibble with 145 rows and 1 variable:
\describe{
  \item{\code{word}}{character vector containing the stopwords}
}}
\source{
Data collected by Finn Årup Nielsen and adapted from https://github.com/fnielsen
}
\usage{
stopwords
}
\description{
A wordlist of Danish stopwords. Based on the 120 most common Danish words
}
\examples{

data(stopwords)
head(stopwords)

}
\keyword{datasets}
